Verify types in custom op schemas
Before this PR, we didn't check that types in a schema were valid. This
is because TorchScript treats unknown types as type variables.

This PR checks types in a schema for the TORCH_LIBRARY APIs. To do this,
we add an `allow_typevars` flag to parseSchema so that TorchScript can
use allow_typevars=True. We also add some error messages for common
mistakes (e.g. using int64_t or double in schema).

Test Plan:
- new tests

ghstack-source-id: 04eac8a9d70d2f82c7e50b1e025985a5f8b188b8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124520